+++ internal/tss/tss.go
@@ -37,7 +37,7 @@ func NewTssService(libp2p *p2p.LibP2PService, dbm *db.DatabaseManager, state *st
 		sigFailChan:    make(chan interface{}, 10),
 		sigTimeoutChan: make(chan interface{}, 10),
 
-		sigMap:                       make(map[string]map[uint64]*signing.LocalParty),
+		sigMap:                       make(map[string]map[int32]*signing.LocalParty),
 		sigRound1P2pMessageMap:       make(map[string]*p2p.Message),
 		sigRound1MessageSendTimesMap: make(map[string]int),
 		sigTimeoutMap:                make(map[string]time.Time),
